Refactor RTF extraction logic from JavaScript to TypeScript
- Converted the RTF extraction implementation to TypeScript, enhancing type safety and clarity.
- Utilized async/await for improved error handling in the extraction process.
- Updated function signatures and streamlined the extraction logic.
- Removed the old JavaScript RTF extractor file.
