--- lib/extractors/rtf.js
@@ -1,89 +1,0 @@
-import { exec } from 'node:child_process';
-import os from 'node:os';
-import path from 'node:path';
-import util from '../util.js';
-import htmlExtract from './html.js';
-
-let types;
-
-/**
- *
- * @param filePath
- * @param options
- * @param cb
- */
-function extractText(filePath, options, cb) {
-  const execOptions = util.createExecOptions('rtf', options);
-  const escapedPath = filePath.replace(/\s/g, '\\ ');
-  // Going to output html from unrtf because unrtf does a great job of
-  // going to html, but does a crap job of going to text. It leaves sections
-  // out, strips apostrophes, leaves nasty quotes in for bullets and more
-  // that I've likely not yet discovered.
-  //
-  // textract can go from html to text on its own, so let unrtf go to html
-  // then extract the text from that
-  //
-  // Also do not have to worry about stripping comments from unrtf text
-  // output since HTML comments are not included in output. Also, the
-  // unrtf --quiet option doesn't work.
-  exec(
-    `unrtf --html --nopict ${escapedPath}`,
-    execOptions,
-    (error, stdout /* , stderr */) => {
-      let err;
-      if (error) {
-        err = new Error(
-          `unrtf read of file named [[ ${path.basename(filePath)} ]] failed: ${
-            error
-          }`,
-        );
-        cb(err, null);
-      } else {
-        htmlExtract.extractFromText(stdout.trim(), {}, cb);
-      }
-    },
-  );
-}
-
-/**
- *
- * @param options
- * @param cb
- */
-function testForBinary(options, cb) {
-  // just non-osx extractor
-  if (os.platform() === 'darwin') {
-    cb(true);
-    return;
-  }
-
-  exec(`unrtf ${__filename}`, (error /* , stdout, stderr */) => {
-    let msg;
-    if (
-      error !== null &&
-      error.message &&
-      error.message.includes('not found')
-    ) {
-      msg =
-        "INFO: 'unrtf' does not appear to be installed, " +
-        'so textract will be unable to extract RTFs.';
-      cb(false, msg);
-    } else {
-      cb(true);
-    }
-  });
-}
-
-// rely on native tools on osx
-if (os.platform() === 'darwin') {
-  types = [];
-  // types = ['application/rtf', 'text/rtf'];
-} else {
-  types = ['application/rtf', 'text/rtf'];
-}
-
-export default {
-  types,
-  extract: extractText,
-  test: testForBinary,
-};
